pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slide-viewer"
version = "1.0.0"
description = "A markdown-powered desktop presentation tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Waleed Abood"},
]
keywords = ["presentation", "markdown", "slides", "PyQt6", "desktop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Multimedia :: Graphics :: Presentation",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyQt6>=6.6,<7",
    "PyQt6-WebEngine>=6.6,<7",
    "markdown>=3.5,<4",
    "pygments>=2.17,<3",
    "openai>=1.0,<3",
    "python-pptx>=0.6.23,<2",
]

[project.urls]
Homepage = "https://github.com/waleedabood/SlideViewer"
Repository = "https://github.com/waleedabood/SlideViewer"
"Bug Tracker" = "https://github.com/waleedabood/SlideViewer/issues"

[project.scripts]
slide-viewer = "slide_viewer.app:main"

[project.gui-scripts]
slide-viewer = "slide_viewer.app:main"

[tool.setuptools.packages.find]
include = ["slide_viewer*"]

[tool.setuptools.package-data]
slide_viewer = ["SLIDE_FORMAT_GUIDE.md"]
